backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.7.4
openai>=2.30.0
pydantic-settings==2.5.2
python-dotenv==1.0.1
langsmith==0.1.117
httpx==0.27.2
python-multipart==0.0.12
pymupdf>=1.25.0
tiktoken>=0.8.0
